+++ CHANGELOG.md
@@ -35,7 +35,7 @@
   });
   ```
 
-- Adds feedback form ([#4320](https://github.com/getsentry/sentry-react-native/pull/4328))
+- Adds feedback form ([#4320](https://github.com/getsentry/sentry-react-native/pull/4328), [#4338](https://github.com/getsentry/sentry-react-native/pull/4338))
 
   You can add the form component in your UI and customise it like:
   ```jsx
@@ -44,6 +44,7 @@
   <FeedbackFormScreen
     {...props}
     closeScreen={props.navigation.goBack}
+    handleChooseFile={ /*If set an attachment button is added to the form*/ }
     styles={{
       submitButton: {
         backgroundColor: '#6a1b9a',